% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundance_dbscan_m_communities.R
\name{abundance_dbscan_m_communities}
\alias{abundance_dbscan_m_communities}
\title{Title 计算由dbscan聚类所产生小群落中的物种丰度}
\usage{
abundance_dbscan_m_communities(entity,minx,miny,maxx,maxy,eps,minPts,edge)
}
\arguments{
\item{entity}{植物个体的分布坐标与物种}

\item{minx}{核心区边界，最小x}

\item{miny}{核心区边界，最小y}

\item{maxx}{核心区边界，最大x}

\item{maxy}{核心区边界，最大y}

\item{eps}{dbscan聚类的eps参数}

\item{minPts}{dbscan聚类的minPts参数}

\item{edge}{基于边界筛选小群落的方法，可选"inside"与"centroid"}
}
\value{
各个小群落中的物种丰度
}
\description{
计算由dbscan聚类所参数小群落中的物种丰度
}
\examples{
entity=M.COMMUNITIES::entity
abundance_dbscan_m_communities(entity,minx=5,miny=5,maxx=45,maxy=45,eps=1,minPts=2,edge="centroid")

}
