% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centroid_hc_m_community.R
\name{centroid_hc_m_community}
\alias{centroid_hc_m_community}
\title{计算由hc聚类所产生小群落的质心}
\usage{
centroid_hc_m_community(entity,minx,miny,maxx,maxy,scale,edge)
}
\arguments{
\item{entity}{植物个体的分布坐标与物种}

\item{minx}{核心区边界，最小x}

\item{miny}{核心区边界，最小y}

\item{maxx}{核心区边界，最大x}

\item{maxy}{核心区边界，最大y}

\item{scale}{hc聚类的尺度}

\item{edge}{基于边界筛选小群落的方法，可选"inside"与"centroid"}
}
\value{
各个小群落的质心坐标
}
\description{
计算由hc聚类所产生小群落的质心
}
\examples{
entity=M.COMMUNITIES::entity
centroid_hc_m_community(entity,minx=5,miny=5,maxx=45,maxy=45,scale=5,edge="inside")
}
