% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CWM_dbscan_m_communities.R
\name{CWM_dbscan_m_communities}
\alias{CWM_dbscan_m_communities}
\title{Title 计算由dbscan聚类所产生小群落的群落平均功能性状值}
\usage{
CWM_dbscan_m_communities(entity,traits,minx,miny,maxx,maxy,eps,minPts,edge)
}
\arguments{
\item{entity}{植物个体的分布坐标与物种}

\item{traits}{物种的功能性状}

\item{minx}{核心区边界，最小x}

\item{miny}{核心区边界，最小y}

\item{maxx}{核心区边界，最大x}

\item{maxy}{核心区边界，最大y}

\item{eps}{dbscan聚类的eps参数}

\item{minPts}{dbscan聚类的minPts参数}

\item{edge}{基于边界筛选小群落的方法，可选"inside"与"centroid"}
}
\value{
小群落的平均功能性状值
}
\description{
计算由dbscan聚类所产生小群落的群落平均功能性状值
}
\examples{
entity=M.COMMUNITIES::entity
traits=M.COMMUNITIES::traits
CWM_dbscan_m_communities(entity,traits,minx=5,miny=5,maxx=45,maxy=45,eps=2,minPts=1,edge="centroid")
}
