% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CWM_hc_m_communities.R
\name{CWM_hc_m_communities}
\alias{CWM_hc_m_communities}
\title{Title 计算由hc聚类所产生小群落的群落平均功能性状值}
\usage{
CWM_hc_m_communities(entity, traits, minx, miny, maxx, maxy, scale, edge)
}
\arguments{
\item{entity}{植物个体的分布坐标与物种}

\item{traits}{物种的功能性状}

\item{minx}{核心区边界，最小x}

\item{miny}{核心区边界，最小y}

\item{maxx}{核心区边界，最大x}

\item{maxy}{核心区边界，最大y}

\item{scale}{hc聚类的scale参数}

\item{edge}{基于边界筛选小群落的方法，可选"inside"与"centroid"}
}
\value{
小群落的平均功能性状值
}
\description{
计算由hc聚类所产生小群落的群落平均功能性状值
}
\examples{
entity=M.COMMUNITIES::entity
traits=M.COMMUNITIES::traits
CWM_hc_m_communities(entity,traits,minx=5,miny=5,maxx=45,maxy=45,scale=5,edge="inside")
}
