% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbscan_m_community_classification.R
\name{dbscan_m_community_classification}
\alias{dbscan_m_community_classification}
\title{Title 基于dbscan聚类对小群落进行划分}
\usage{
dbscan_m_community_classification(entity,minx,miny,maxx,maxy,eps,minPts,edge)
}
\arguments{
\item{entity}{植物个体的分布坐标与物种}

\item{minx}{核心区边界，最小x}

\item{miny}{核心区边界，最小y}

\item{maxx}{核心区边界，最大x}

\item{maxy}{核心区边界，最大y}

\item{eps}{dbscan聚类的eps参数}

\item{minPts}{dbscan聚类的minPts参数}

\item{edge}{基于边界筛选小群落的方法，可选"inside"与"centroid"}
}
\value{
小群落的划分
}
\description{
基于dbscan聚类对小群落进行划分
}
\examples{
entity=M.COMMUNITIES::entity
dbscan_m_community_classification(entity,minx=5,miny=5,maxx=5,maxy=5,eps=1,minPts=2,edge="centroid")
}
