% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc_m_community_Beta_MPD_pair.R
\name{hc_m_community_Beta_MPD_pair}
\alias{hc_m_community_Beta_MPD_pair}
\title{Title 计算由hc聚类所产生小群落间的平均配对功能性状距离(SES.贝塔MPD)}
\usage{
hc_m_community_Beta_MPD_pair(entity,traits,minx,miny,maxx,maxy,scale,replicate,edge)
}
\arguments{
\item{entity}{植物个体的分布坐标与物种}

\item{traits}{物种的功能性状}

\item{minx}{核心区边界，最小x}

\item{miny}{核心区边界，最小y}

\item{maxx}{核心区边界，最大x}

\item{maxy}{核心区边界，最大y}

\item{scale}{hc聚类的scale参数}

\item{replicate}{零模型重复次数}

\item{edge}{基于边界筛选小群落的方法，可选"inside"与"centroid"}
}
\value{
SES.贝塔MPD
}
\description{
计算由hc聚类所产生小群落间的平均配对功能性状距离(SES.贝塔MPD)
}
\examples{
entity=M.COMMUNITIES::entity
traits=M.COMMUNITIES::traits
hc_m_community_Beta_MPD_pair(entity,traits,minx=5,miny=5,maxx=45,maxy=45,scale=5,replicate=99,edge="inside")
}
