% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc_m_community_Bray_Curtis_pair.R
\name{hc_m_community_Bray_Curtis_pair}
\alias{hc_m_community_Bray_Curtis_pair}
\title{计算小群落物种组成的Bray_Curtis距离}
\usage{
hc_m_community_Bray_Curtis_pair(entity,minx,miny,maxx,maxy,scale,edge)
}
\arguments{
\item{entity}{植物个体的分布坐标与物种}

\item{minx}{核心区边界，最小x}

\item{miny}{核心区边界，最小y}

\item{maxx}{核心区边界，最大x}

\item{maxy}{核心区边界，最大y}

\item{scale}{hc聚类的scale参数}

\item{edge}{基于边界筛选小群落的方法，可选"inside"与"centroid"}
}
\value{
小群落间的Bray_Curtis距离
}
\description{
计算小群落物种组成的Bray_Curtis距离
}
\examples{
entity=M.COMMUNITIES::entity
hc_m_community_Bray_Curtis_pair(entity,minx=5,miny=5,maxx=45,maxy=45,scale=5,edge="centroid")
}
