% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc_m_community_classification.R
\name{hc_m_community_classification}
\alias{hc_m_community_classification}
\title{Title 基于hc聚类对小群落进行划分}
\usage{
hc_m_community_classification(entity,minx,miny,maxx,maxy,scale,edge)
}
\arguments{
\item{entity}{植物个体的分布坐标与物种}

\item{minx}{核心区边界，最小x}

\item{miny}{核心区边界，最小y}

\item{maxx}{核心区边界，最大x}

\item{maxy}{核心区边界，最大y}

\item{scale}{hc聚类的scale参数}

\item{edge}{基于边界筛选小群落的方法，可选"inside"与"centroid"}
}
\value{
小群落的划分
}
\description{
基于hc聚类对小群落进行划分
}
\examples{
entity=M.COMMUNITIES::entity
hc_m_community_classification(entity,minx=5,miny=5,maxx=45,maxy=45,scale=5,edge="centroid")
}
