% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc_m_community_scale_select.R
\name{hc_m_community_scale_select}
\alias{hc_m_community_scale_select}
\title{Title 基于点格局分析的方法，判定以hc聚类划分小群落时的合理聚类尺度}
\usage{
hc_m_community_scale_select(entity,minx,miny,maxx,maxy,window,two_nested_clu=FALSE)
}
\arguments{
\item{entity}{植物个体的分布坐标与物种}

\item{minx}{研究范围的最小x}

\item{miny}{研究范围的最小y}

\item{maxx}{研究范围的最大x}

\item{maxy}{研究范围的最大y}

\item{window}{点格局分析时的窗口大小}

\item{two_nested_clu}{逻辑变量，可为T或F，判定是否模拟双尺度镶嵌托马斯过程}
}
\value{
聚类尺度
}
\description{
基于点格局分析的方法，判定以hc聚类划分小群落时的合理聚类尺度
}
\examples{
entity=M.COMMUNITIES::entity
hc_m_community_scale_select(entity,minx=0,miny=0,maxx=50,maxy=50,window=35,two_nested_clu=T)
}
