% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hc_m_community_classification.R
\name{plot_hc_m_community_classification}
\alias{plot_hc_m_community_classification}
\title{绘制基于hc聚类划分的小群落的分布图}
\usage{
plot_hc_m_community_classification(entity,minx,miny,maxx,maxy,scale,edge)
}
\arguments{
\item{entity}{植物个体的分布坐标与物种}

\item{minx}{核心区边界，最小x}

\item{miny}{核心区边界，最小y}

\item{maxx}{核心区边界，最大x}

\item{maxy}{核心区边界，最大y}

\item{scale}{hc聚类的scale参数}

\item{edge}{基于边界筛选小群落的方法，可选"inside"与"centroid"}
}
\value{
小群落空间分布图
}
\description{
绘制基于hc聚类划分的小群落的分布图
}
\examples{
entity=M.COMMUNITIES::entity
plot_hc_m_community_classification(entity,minx=5,miny=5,maxx=45,maxy=45,scale=5,edge="inside")
}
