% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdiversity_dbscan_m_communities.R
\name{sdiversity_dbscan_m_communities}
\alias{sdiversity_dbscan_m_communities}
\title{计算基于dbscan聚类所划分小群落的物种多样性指数}
\usage{
sdiversity_dbscan_m_communities(entity,minx,miny,maxx,maxy,eps,minPts,edge)
}
\arguments{
\item{entity}{植物个体的分布坐标与物种}

\item{minx}{核心区边界，最小x}

\item{miny}{核心区边界，最小y}

\item{maxx}{核心区边界，最大x}

\item{maxy}{核心区边界，最大y}

\item{eps}{dbscan聚类的eps参数}

\item{minPts}{dbscan聚类的minPts参数}

\item{edge}{基于边界筛选小群落的方法，可选"inside"与"centroid"}
}
\value{
物种多样性指数
}
\description{
计算基于dbscan划分小群落的物种多样性指数
}
\examples{
entity=M.COMMUNITIES::entity
sdiversity_dbscan_m_communities(entity,minx=5,miny=5,maxx=45,maxy=45,eps=2,minPts=1,edge="centroid")
}
