% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdiversity_hc_m_communities.R
\name{sdiversity_hc_m_communities}
\alias{sdiversity_hc_m_communities}
\title{计算基于hc聚类所划分小群落的物种多样性指数}
\usage{
sdiversity_hc_m_communities(entity,minx,miny,maxx,maxy,scale,edge)
}
\arguments{
\item{entity}{植物个体的分布坐标与物种}

\item{minx}{核心区边界，最小x}

\item{miny}{核心区边界，最小y}

\item{maxx}{核心区边界，最大x}

\item{maxy}{核心区边界，最大y}

\item{scale}{hc聚类的尺度}

\item{edge}{基于边界筛选小群落的方法，可选"inside"与"centroid"}
}
\value{
物种多样性指数
}
\description{
计算基于hc聚类所划分小群落的物种多样性指数
}
\examples{
entity=M.COMMUNITIES::entity
sdiversity_hc_m_communities(entity,minx=5,miny=5,maxx=45,maxy=45,scale=5,edge="centroid")

}
