% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VPA_priority_enviroment_hc.R
\name{VPA_priority_enviroment_hc}
\alias{VPA_priority_enviroment_hc}
\title{以基于物种存在的SES.MPD及SES.E对基于物种丰度的SES.MPD进行分解}
\usage{
VPA_priority_enviroment_hc(entity,traits,minx,miny,maxx,maxy,scale,replicate,edge="centroid")
}
\arguments{
\item{entity}{植物个体的分布坐标与物种}

\item{traits}{物种的功能性状}

\item{minx}{核心区边界，最小x}

\item{miny}{核心区边界，最小y}

\item{maxx}{核心区边界，最大x}

\item{maxy}{核心区边界，最大y}

\item{scale}{hc聚类的尺度}

\item{replicate}{零模型重复次数}

\item{edge}{基于边界筛选小群落的方法，可选"inside"与"centroid"}
}
\value{
基于物种丰度(存在)的SES.MPD、SES.E
}
\description{
以基于物种存在的SES.MPD及SES.E对基于物种丰度的SES.MPD进行分解
}
\examples{
entity=M.COMMUNITIES::entity
traits=M.COMMUNITIES::traits
VPA_priority_enviroment_hc(entity,traits,minx=5,miny=5,maxx=45,maxy=45,scale=5,replicate=99,edge="centroid")
}
